requirements.txt
APScheduler==3.6.3
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
cachetools==4.2.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
fake-useragent==0.1.11
h11==0.13.0
idna==3.3
lxml==4.9.0
numpy==1.22.4
outcome==1.2.0
pandas==1.4.2
psycopg2==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.0
schedule==1.1.0
scrapper-boilerplate==0.2.3
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.37
tornado==6.1
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
webdriver-manager==3.7.0
wsproto==1.1.0
